feat: メソッドチェーン機能と統一ユーティリティを実装し、画像処理機能を拡張
- すべての画像処理トレイトに対してメソッドチェーン機能をサポート

  - AlphaPremultiply: インプレイス処理用バリアントと所有型メソッドを追加
  - ApplyAlphaMask: インプレイスアルファ操作用のModifyAlphaトレイトを追加
  - NLMeans: インプレイスノイズ除去用のnl_means_mutメソッドを追加
  - Padding: 次元処理を適切に扱うようにAPIを改善
  - ForegroundEstimator/ClipMinimumBorder: 隠し_mutバリアントを追加

- 画像統一処理用ユーティリティを導入

  - unify_rgb_images関数とunify_gray_images関数を追加
  - 混合精度演算時の自動型昇格をサポート
  - コンパイル時型解決用のLargerTypeトレイトを追加

- テストカバレッジと型整合性を向上

  - テストコードを適切な型アサーションに更新
  - 統一処理関数に対する包括的なテストを追加
  - パディング処理テストにおける位置型の扱いを修正

- READMEを更新し、NL-Meansノイズ除去機能のドキュメントを追加

+++ tests/property_tests.rs
@@ -15,6 +15,7 @@ fn image_dimensions() -> impl Strategy<Value = (u32, u32)> {
 }
 
 /// Strategy for generating larger image dimensions for less expensive tests
+#[allow(dead_code)]
 fn larger_image_dimensions() -> impl Strategy<Value = (u32, u32)> {
     (10u32..=50, 10u32..=50)
 }
@@ -65,6 +66,7 @@ fn create_test_rgb_image_with_pattern(
 }
 
 /// Create a test RGBA image with given dimensions and fill pattern
+#[allow(dead_code)]
 fn create_test_rgba_image_with_pattern(
     width: u32,
     height: u32,
@@ -298,8 +300,8 @@ proptest! {
         let image = create_test_rgb_image_with_pattern(orig_width, orig_height, |_, _| Rgb([0, 0, 0]));
 
         if let Ok((_, (pos_x, pos_y))) = image.add_padding((pad_width, pad_height), position, Rgb([255, 255, 255])) {
-            prop_assert!(pos_x + orig_width <= pad_width);
-            prop_assert!(pos_y + orig_height <= pad_height);
+            prop_assert!(pos_x + orig_width as i64 <= pad_width as i64);
+            prop_assert!(pos_y + orig_height as i64 <= pad_height as i64);
         }
     }
